#2537 Added InfluxDB and Grafana to Docker-compose. There's now a framework to log runtimes of scheduled tasks to Grafana. Not all tasks use them yet but the dungeonroute: ones do for now. Will do the rest tomorrow.

+++ app/Console/Commands/Scheduler/Telemetry/Measurement/MySqlStats.php
@@ -12,8 +12,8 @@ class MySqlStats extends Measurement
      */
     public function getPoints(): array
     {
-        $threadsConnected = DB::select('SHOW STATUS WHERE `variable_name` = "Threads_connected"')[0]->Value;
-        $maxThreads       = DB::select('SHOW VARIABLES LIKE "max_connections"')[0]->Value;
+        $threadsConnected = (int)DB::select('SHOW STATUS WHERE `variable_name` = "Threads_connected"')[0]->Value;
+        $maxThreads       = (int)DB::select('SHOW VARIABLES LIKE "max_connections"')[0]->Value;
 
         return [
             new Point(
